+++ libraries/TTBAS_LIB/tTVscreen.h
@@ -80,13 +80,13 @@ class tTVscreen : public tscreenBase, public tGraphicDev {
     void Serial_Ctrl(int16_t ch);
     void reset_kbd(uint8_t kbd_type=false);
     void putch(uint8_t c);                            // 文字の出力
-    uint8_t get_ch();                                 // 文字の取得
-    inline uint8_t tryGetChar() {
+    uint16_t get_ch();                                 // 文字の取得
+    inline uint16_t tryGetChar() {
       return ps2read();
     }
     inline uint8_t getDevice() {return dev;};         // 文字入力元デバイス種別の取得
     bool isKeyIn();                                // キー入力チェック 
-    inline uint8_t peekKey() {
+    inline uint16_t peekKey() {
       return ps2peek();
     }
     uint8_t edit();                                   // スクリーン編集
